[cleanup] rename controllers to routes
